refactor: code convention
