--- src/main/java/techcourse/myblog/HelloWorldController.java
@@ -1,22 +1,0 @@
-package techcourse.myblog;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-@Controller
-public class HelloWorldController {
-    @GetMapping("/helloworld")
-    @ResponseBody
-    public String blogNameByGet(String blogName) {
-        return blogName;
-    }
-
-    @PostMapping("/helloworld")
-    @ResponseBody
-    public String blogNameByPost(@RequestBody String blogName) {
-        return blogName;
-    }
-}
