--- src/main/java/techcourse/myblog/domain/InvalidArticleException.java
@@ -1,7 +1,0 @@
-package techcourse.myblog.domain;
-
-public class InvalidArticleException extends RuntimeException {
-    public InvalidArticleException(String msg) {
-        super(msg);
-    }
-}
