+++ src/main/java/techcourse/myblog/dto/UserInfo.java
@@ -1,4 +1,4 @@
-package techcourse.myblog;
+package techcourse.myblog.dto;
 
 public interface UserInfo {
 }
