+++ src/main/resources/application.properties
@@ -7,11 +7,11 @@ spring.resources.static-locations=classpath:/static/
 spring.resources.add-mappings=true 
 
 # Datasource 설정
-spring.jpa.hibernate.ddl-auto=update
-#spring.jpa.hibernate.ddl-auto=create
+#spring.jpa.hibernate.ddl-auto=update
+spring.jpa.hibernate.ddl-auto=create
 spring.datasource.url=jdbc:mysql://localhost:3306/myblog?useUnicode=true&characterEncoding=utf8&allowPublicKeyRetrieval=true&useSSL=false&serverTimezone=UTC
-spring.datasource.username=bloguser
-spring.datasource.password=bloguser
+spring.datasource.username=luffy
+spring.datasource.password=159456
 spring.jpa.database-platform=org.hibernate.dialect.MySQL5InnoDBDialect
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
