--- src/test/java/techcourse/myblog/MyblogApplicationTests.java
@@ -1,16 +1,0 @@
-package techcourse.myblog;
-
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-
-@ExtendWith(SpringExtension.class)
-@SpringBootTest
-public class MyblogApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
